Lvup: Refactoring
